code securing for edit and delete

+++ delete.php
@@ -5,13 +5,30 @@
      notLogined();
 ?>
 <?php 
-    $sql = "DELETE FROM to_do_list WHERE id = :id";
+    // old data
+    $sql = "SELECT * FROM to_do_list WHERE id = :id";
     $stmt = $conn->prepare($sql);
     $stmt->execute([
         ":id" => $_GET['id'],
     ]);
+    $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
-    setcookie("success", "You deleted a list");
-    header("Location: ./dashboard.php");
+    $currentUserId = getUserID();
+
+    
+    // to prevent to deleting other users data 
+   if($row['user_id'] != $currentUserId){
+        block($row['user_id'],$currentUserId);
+   } else {
+        $sql = "DELETE FROM to_do_list WHERE id = :id";
+        $stmt = $conn->prepare($sql);
+        $stmt->execute([
+            ":id" => $_GET['id'],
+        ]);
+
+        setcookie("success", "You deleted a list");
+        header("Location: ./dashboard.php");
+   }
 ?>
+
